Wael

+++ database/migrations/2024_12_06_130748_create_questions_table.php
@@ -11,7 +11,7 @@ class CreateQuestionsTable extends Migration {
 		Schema::create('questions', function(Blueprint $table) {
 			$table->id();
 			$table->string('question');
-			$table->unsignedBigInteger('test_id')->unsigned();
+			$table->integer('test_id')->unsigned();
             $table->timestamps();
         });
 	}
